requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
SQLAlchemy==2.0.12
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
alembic==1.10.4
asyncpg==0.27.0